+++ CHANGELOG.md
@@ -25,6 +25,9 @@ Language
 
   This feature is experimental.
 
+* The options `--subtyping` and `--no-subtyping` have been removed
+  (see [#5427](https://github.com/agda/agda/issues/5427)).
+
 * The cubical interval `I` now belongs to its own sort, `IUniv`, rather
   than `SSet`. For `J : ISet` and `A : J → Set l`, we have
   `(j : J) → A : Set l`, that is, the type of functions from a type in `ISet`